backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
pydantic==2.7.4
psycopg[binary]>=3.1.19
asyncpg==0.29.0
redis==5.0.4
qdrant-client==1.9.1
optimum-intel==1.17.0
openvino>=2024.1.0
sentence-transformers==3.0.1
aiohttp==3.9.5
beautifulsoup4==4.12.3
duckduckgo-search>=6.1.7
alembic==1.13.1
